feat(support)!: move env/APPIUM_HOME resolution into appium package
`env` (resolveAppiumHome/hasAppiumDependency/resolveManifestPath) and the
`normalize-package-data`-based package.json reader it relied on were only
ever consumed by the `appium` package (directly, or via `npm.ts`'s
`installPackage`, itself only called from `appium`). Move them there so
`@appium/support` no longer needs `normalize-package-data` at all.

`npm.installPackage` now takes a `hasAppiumDependency` boolean computed by
the caller instead of importing the check itself, avoiding a
support -> appium dependency. The `env` export is removed from
`@appium/support`.
