+++ packages/appium/package.json
@@ -79,6 +79,7 @@
     "axios": "1.19.0",
     "lilconfig": "3.1.3",
     "lru-cache": "11.5.2",
+    "normalize-package-data": "8.0.0",
     "ora": "5.4.1",
     "semver": "7.8.5",
     "teen_process": "4.2.1",
