+++ packages/appium/test/e2e/fixture/appium-v1-dependency.package.json
