+++ packages/appium/test/e2e/fixture/appium-v1-package/package.json
