+++ packages/appium/test/e2e/fixture/appium-v2-dependency.package.json
