+++ packages/appium/test/e2e/fixture/appium-v2-package/package.json
