+++ packages/support/lib/internal/index.ts
@@ -2,12 +2,3 @@
  * Package-internal utilities. Not exported from `@appium/support`.
  */
 export {createBase64EncodeStream} from './base64-encode-stream.js';
-export {
-  type NormalizedPackageJson,
-  type NormalizeOptions,
-  packageDirectorySync,
-  type PackageJson,
-  readPackage,
-  type ReadPackageOptions,
-  readPackageSync,
-} from './read-package.js';
