+++ packages/support/package.json
@@ -61,7 +61,6 @@
     "glob": "13.0.6",
     "klaw": "4.1.0",
     "lockfile": "1.0.4",
-    "normalize-package-data": "8.0.0",
     "plist": "5.0.0",
     "pluralize": "8.0.0",
     "sanitize-filename": "1.6.4",
